+++ packages/terra-alert/CHANGELOG.md
@@ -5,7 +5,7 @@
 ## 4.69.0 - (June 14, 2023)
 
 * Added
-  * Added focus shift to the rendered notification banners content when notification is an alert and includes an action element
+  * Added focus shift to rendered `alert` notification banners content and includes an action element and/or dismiss.
 
 * Changed
   * Locked `uuid` to `8.2.0` to maintain IE compatibility.